refactor: move MatchingPairs and PlayCard to own directory; move visual logic
