--- frontend/src/components/Playback/MatchingPairs.scss
@@ -1,119 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Rajdhani&display=swap');
-
-@media (min-aspect-ratio: 1/1) {
-    $card: 16vh;
-}
-
-@media (max-aspect-ratio: 1/1) {
-    $card: 18vw;
-}
-
-.aha__page .container {
-    padding-top: 2vh;
-}
-
-.aha__matching-pairs {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-}
-
-@media (min-aspect-ratio: 1/1)  {
-    .playing-board {
-        grid-template-columns: 16vh 16vh 16vh 16vh;
-        column-gap: 1.5vh;
-        row-gap: 1.5vh;
-
-        &.playing-board--three-columns {
-            grid-template-columns: 16vh 16vh 16vh;
-        }
-    }
-}
-
-@media (max-aspect-ratio: 1/1)  {
-    .playing-board {
-        grid-template-columns: 18vw 18vw 18vw 18vw;
-        column-gap: 2vw;
-        row-gap: 2vw;
-
-        &.playing-board--three-columns {
-            grid-template-columns: 18vw 18vw 18vw;
-        }
-    }
-}
-
-.playing-board {
-    display: inline-grid;
-}
-
-.matching-pairs__score-feedback {
-
-    font-weight: 700;
-
-    &--small-bottom-right {
-        position: fixed;
-        justify-content: flex-end;
-        bottom: 0;
-        right: 0;
-        text-align: right;
-        font-size: 0.5rem;
-        backface-visibility: hidden;
-        font-weight: normal;
-
-        .matching-pairs__feedback, .matching-pairs__score {
-            float: none;
-        }
-    }
-}
-
-.matching-pairs__feedback, .matching-pairs__score {
-    font-family: 'Rajdhani', sans-serif;
-    line-height: 1;
-    text-align: center;
-    padding-bottom: .7rem;
-    font-size: 95%;
-    width: auto;
-
-    @media(min-width: 900px) {
-        font-size: 150%;
-    }
-}
-
-.matching-pairs__feedback {
-    float: left;
-    color: white;
-
-    &.fbnomatch {
-        color: $gray;
-    }
-    &.fblucky {
-        color: $yellow;
-    }
-    &.fbmemory {
-        color: $teal;
-    }
-    &.fbmisremembered {
-        color: $red;
-    }
-    &.nomessage {
-        opacity: 0;
-    }
-} 
-
-.matching-pairs__score {
-    float: right;
-
-    &.noscore {
-        opacity: 0;
-    }
-}
-
-.turnscore {
-    position: fixed;
-    bottom: -2rem;
-    left: -2rem;
-    &.noturnscore {
-        opacity: 0;
-    }
-}
