+++ frontend/src/components/Playback/Playback.js
@@ -8,9 +8,8 @@ import AutoPlay from "./Autoplay";
 import PlayButton from "../PlayButton/PlayButton";
 import MultiPlayer from "./MultiPlayer";
 import ImagePlayer from "./ImagePlayer";
-import MatchingPairs from "./MatchingPairs";
+import MatchingPairs from "../MatchingPairs/MatchingPairs";
 import Preload from "../Preload/Preload";
-import VisualMatchingPairs from "components/VisualMatchingPairs/VisualMatchingPairs";
 
 export const AUTOPLAY = "AUTOPLAY";
 export const BUTTON = "BUTTON";
@@ -18,7 +17,6 @@ export const MULTIPLAYER = "MULTIPLAYER";
 export const IMAGE = "IMAGE";
 export const MATCHINGPAIRS = "MATCHINGPAIRS";
 export const PRELOAD = "PRELOAD";
-export const VISUALMATCHINGPAIRS = "VISUALMATCHINGPAIRS";
 
 const Playback = ({
     playbackArgs,
@@ -218,13 +216,6 @@ const Playback = ({
                         {...attrs}
                         stopAudioAfter={playbackArgs.stop_audio_after}
                         showAnimation={playbackArgs.show_animation}
-                        scoreFeedbackDisplay={playbackArgs.score_feedback_display}
-                    />
-                );
-            case VISUALMATCHINGPAIRS:
-                return (
-                    <VisualMatchingPairs
-                        {...attrs}
                         scoreFeedbackDisplay={playbackArgs.score_feedback_display}
                     />
                 );
